+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/workshops/ModifiedWorkshopDto.kt
@@ -1,7 +1,0 @@
-package com.app.arcabyolimpo.data.remote.dto.workshops
-
-data class ModifyWorkshopDto(
-    val success: Boolean,
-    val message: String,
-    val idTaller: String?
-)