+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.0] - 19/11/2025
+### Added
+- Added modify for workshops.
+
 ## [0.1.0] - 18/11/2025
 
 ### Fixed